requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
spotipy==2.23.0
python-dotenv==1.0.0 